[CEDS-3091] Custom dns for first contact file upload
The first file in a batch upload is always the contacts file created and uploaded by the SFUS frontend itself.

In non-E2E environments (like development, staging and ET) the SFUS frontend exposed a test-only endpoint to mock the S3 bucket url that files are normally upload to. So in effect for this first file the SFUS frontend is calling itself!

The domain names that a user can access from their browser and the the service can access from its MDTP environment are not the same. So to facilitate the testing of the SFUS frontend in non-E2E environments the stub has this special behaviour of switching the domain name for uploading the first contacts file to an internal domain that the SFUS frontend can reach.